album

+++ src/main/java/online/kheops/auth_server/resource/AlbumRessource.java
@@ -1,9 +1,9 @@
 package online.kheops.auth_server.resource;
 
-import online.kheops.auth_server.KheopsPrincipal;
 import online.kheops.auth_server.KheopsPrincipalInterface;
 import online.kheops.auth_server.album.*;
 import online.kheops.auth_server.annotation.Secured;
+import online.kheops.auth_server.capability.ScopeType;
 import online.kheops.auth_server.user.UserNotFoundException;
 import online.kheops.auth_server.user.UsersPermission;
 import online.kheops.auth_server.util.PairListXTotalCount;
@@ -33,7 +33,13 @@ public class AlbumRessource {
                              @FormParam("addSeries") Boolean addSeries, @FormParam("writeComments") Boolean writeComments,
                              @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasUserWriteAccess()) {
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
+
         final UsersPermission usersPermission = new UsersPermission();
 
         if (addUser != null) { usersPermission.setAddUser(addUser); }
@@ -65,8 +71,13 @@ public class AlbumRessource {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     public Response getAlbums(@Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
         final PairListXTotalCount<AlbumResponses.AlbumResponse> pairAlbumsTotalAlbum;
+
+        if(!kheopsPrincipal.hasUserReadAccess()){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             pairAlbumsTotalAlbum = Albums.getAlbumList(callingUserPk, uriInfo.getQueryParameters());
@@ -92,8 +103,13 @@ public class AlbumRessource {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getAlbum(@PathParam("album") Long albumPk,@Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
         final AlbumResponses.AlbumResponse albumResponse;
+
+        if(!kheopsPrincipal.hasAlbumAccess(albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             albumResponse = Albums.getAlbum(callingUserPk, albumPk);
@@ -107,7 +123,6 @@ public class AlbumRessource {
 
         return Response.status(Response.Status.OK).entity(albumResponse).build();
     }
-
 
     @PATCH
     @Secured
@@ -123,7 +138,13 @@ public class AlbumRessource {
                               @FormParam("notificationNewComment") Boolean notificationNewComment,
                               @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasAlbumPermission(UsersPermission.UsersPermissionEnum.EDIT_ALBUM, albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
+
         final AlbumResponses.AlbumResponse albumResponse;
         final UsersPermission usersPermission = new UsersPermission();
 
@@ -156,7 +177,12 @@ public class AlbumRessource {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     public Response deleteAlbum(@PathParam("album") Long albumPk,@Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasAlbumPermission(UsersPermission.UsersPermissionEnum.DELETE_ALBUM, albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             Albums.deleteAlbum(callingUserPk, albumPk);
@@ -178,7 +204,13 @@ public class AlbumRessource {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getUsersAlbum(@PathParam("album") Long albumPk,@Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasAlbumPermission(UsersPermission.UsersPermissionEnum.LIST_USERS, albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
+
         final List<AlbumResponses.UserAlbumResponse> usersAlbumResponse;
 
         try {
@@ -202,7 +234,12 @@ public class AlbumRessource {
     public Response addUser(@PathParam("album") Long albumPk, @PathParam("user") String user,
                             @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasAlbumPermission(UsersPermission.UsersPermissionEnum.ADD_USER, albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             Albums.addUser(callingUserPk, user, albumPk, false);
@@ -224,7 +261,12 @@ public class AlbumRessource {
     public Response addAdmin(@PathParam("album") Long albumPk, @PathParam("user") String user,
                              @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasAlbumPermission(UsersPermission.UsersPermissionEnum.ADD_ADMIN, albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             Albums.addUser(callingUserPk, user, albumPk, true);
@@ -246,7 +288,12 @@ public class AlbumRessource {
     public Response removeAdmin(@PathParam("album") Long albumPk, @PathParam("user") String user,
                                 @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasAlbumPermission(UsersPermission.UsersPermissionEnum.REMOVE_ADMIN, albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             Albums.removeAdmin(callingUserPk, user, albumPk);
@@ -268,7 +315,12 @@ public class AlbumRessource {
     public Response deleteUser(@PathParam("album") Long albumPk, @PathParam("user") String user,
                                @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!kheopsPrincipal.hasAlbumPermission(UsersPermission.UsersPermissionEnum.REMOVE_USER, albumPk)){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             Albums.deleteUser(callingUserPk, user, albumPk);
@@ -289,7 +341,12 @@ public class AlbumRessource {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     public Response addFavorites(@PathParam("album") Long albumPk, @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!(kheopsPrincipal.hasUserWriteAccess() && kheopsPrincipal.hasAlbumAccess(albumPk))){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             Albums.setFavorites(callingUserPk, albumPk, true);
@@ -307,7 +364,12 @@ public class AlbumRessource {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     public Response deleteFavorites(@PathParam("album") Long albumPk, @Context SecurityContext securityContext) {
 
-        final long callingUserPk = ((KheopsPrincipalInterface)securityContext.getUserPrincipal()).getDBID();
+        KheopsPrincipalInterface kheopsPrincipal = ((KheopsPrincipalInterface)securityContext.getUserPrincipal());
+        final long callingUserPk = kheopsPrincipal.getDBID();
+
+        if(!(kheopsPrincipal.hasUserWriteAccess() && kheopsPrincipal.hasAlbumAccess(albumPk))){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         try {
             Albums.setFavorites(callingUserPk, albumPk, false);
